+++ pages/index.jsx
@@ -4,48 +4,55 @@ import getConfig from 'next/config';
 
 import '@material/layout-grid/dist/mdc.layout-grid.css';
 
+import ItemData from '../data/itemData';
 import ItemTitle from '../components/itemTitle';
 import ItemPrice from '../components/itemPrice';
 import ItemImages from '../components/itemImages'; // ItemImage
-import ItemCallouts from '../components/itemCallouts';
-import ItemPurchase from '../components/itemPurchase'; // ItemQuantity, ItemPickupButton, and ItemCartButton
-import ItemDisclosures from '../components/itemDisclosures';
-import ItemActions from '../components/itemActions'; // ItemActionButton
+import ItemPromos from '../components/itemPromos';
+import ItemPurchase from '../components/itemPurchase'; // ItemButton
+import ItemDisclosure from '../components/itemDisclosure';
+import ItemActions from '../components/itemActions'; // ItemButton
 import ItemReviews from '../components/itemReviews'; // ItemTopPro and ItemTopCon
 import ItemHighlights from '../components/itemHighlights';
 
 const { publicRuntimeConfig } = getConfig();
 
 export default class Index extends React.Component {
-  static async getInitialProps({ req }) {
-    const item = await fetch(`${publicRuntimeConfig.hostname}/data.json`)
-      .then(response => response.json())
-      .then(data => data.CatalogEntryView[0]);
-    return { item };
+  static async getInitialProps() {
+    const item = new ItemData();
+    return item.get();
   }
 
   render() {
-    const { item } = this.props;
-    console.log(item);
+    const {
+      title,
+      images,
+      price,
+      promos,
+      availability,
+      disclosure,
+    } = this.props;
+    console.log(title, images, price, promos, availability);
 
     return (
       <div className="mdc-top-app-bar--fixed-adjust">
         <div className="mdc-layout-grid">
           <div className="mdc-layout-grid__inner">
             <div className="mdc-layout-grid__cell--span-6-desktop mdc-layout-grid__cell--span-4-phone">
-              <ItemTitle title={item.title} />
-              <ItemImages
-                primaryImage={item.Images[0].PrimaryImage[0].image}
-                alternateImages={item.Images[0].AlternateImages}
-              />
+              <ItemTitle title={title} />
+              <ItemImages images={images} />
             </div>
             <div className="mdc-layout-grid__cell--span-6-desktop mdc-layout-grid__cell--span-4-phone">
-              <ItemPrice
-                price={item.Offers[0].OfferPrice[0].formattedPriceValue}
+              <ItemPrice price={price} />
+              <ItemPromos promos={promos} />
+              <ItemPurchase
+                inStore={availability.inStore}
+                online={availability.online}
               />
-              <ItemCallouts callouts={item.Promotions} />
-              <ItemPurchase availability={item.purchasingChannelCode} />
-              {/* <ItemDisclosures returnPolicy={item.ReturnPolicy[0].legalCopy} /> */}
+              <ItemDisclosure
+                title={disclosure.title}
+                description={disclosure.description}
+              />
               {/* <ItemActions /> */}
             </div>
           </div>